:minibus: Rename files a bit
