+++ lib/main.js
@@ -1,49 +1,45 @@
 'use babel'
 
 import {CompositeDisposable} from 'atom'
-import {LinterUI} from './provider'
+import {Panel} from './panel'
 
-export default {
-  config: {
-    showPanel: {
-      type: 'boolean',
-      description: 'Show panel on the bottom',
-      default: true
-    },
-    showIssuesFromAllFiles: {
-      title: 'Show issues from all files',
-      description: 'Display all of the issues in bottom panel instead of only current file ones',
-      type: 'boolean',
-      default: false
-    },
-    gutterEnabled: {
-      title: 'Highlight Error Lines in Gutter',
-      type: 'boolean',
-      default: true
-    },
-    gutterPosition: {
-      title: 'Position of Gutter Highlights',
-      description: 'Relative to line number gutter',
-      type: 'string',
-      enum: ['Left', 'Right'],
-      default: 'Right'
-    },
-    underlineIssues: {
-      title: 'Underline issues',
-      type: 'boolean',
-      default: true
-    }
-  },
-  activate: function() {
-    this.ui = new LinterUI()
+export class LinterUI {
+  constructor() {
+    this.name = 'Linter'
+    this.editors = new Set()
+    this.editorRegistry = null
+    this.panel = null
     this.subscriptions = new CompositeDisposable()
+  }
+  activate(editorRegistry) {
+    this.editorRegistry = editorRegistry
 
-    this.subscriptions.add(this.ui)
-  },
-  deactivate: function() {
+    this.subscriptions.add(atom.config.observe('linter-ui-default.showPanel', showPanel => {
+      if (showPanel && !this.panel) {
+        this.panel = new Panel()
+        this.panel.onDidDestroy(() => {
+          atom.config.set('linter-ui-default.showPanel', false)
+        })
+      } else if (!showPanel && this.panel) {
+        this.panel.dispose()
+        this.panel = null
+      }
+    }))
+  }
+  didCalculateMessages({added, removed, messages}) {
+
+  }
+  didBeginLinting(linter, filePath) {
+
+  }
+  didFinishLinting(linter, filePath) {
+
+  }
+  dispose() {
+    this.editorRegistry = null
+    if (this.panel) {
+      this.panel.dispose()
+    }
     this.subscriptions.dispose()
-  },
-  provideUI: function() {
-    return this.ui
   }
 }
